replace UpdateContract and UpdateService
